backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
aiosqlite
asyncpg
pydantic
pydantic-settings
python-dotenv
email-validator
greenlet
httpx
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt
argon2-cffi
langchain
langchain-google-genai
langchain-community
chromadb
